util/fossilize_db: Use uint64_t for file size.
For those 32-bit systems with 4G of cache.

Fixes: 2ec1bff0f3a "util/fossilize_db: Split out reading the index."
Part-of: <https://gitlab.freedesktop.org/mesa/mesa/-/merge_requests/12266>

+++ src/util/fossilize_db.c
@@ -112,7 +112,7 @@ update_foz_index(struct foz_db *foz_db, FILE *db_idx, unsigned file_idx)
 {
    uint64_t offset = ftell(db_idx);
    fseek(db_idx, 0, SEEK_END);
-   size_t len = ftell(db_idx);
+   uint64_t len = ftell(db_idx);
    uint64_t parsed_offset = offset;
 
    if (offset == len)
